split javascript into modules
